removed FrameType and Enter
This also removes the following passes as these optimization are highly
dependent on the Enter primop:
* dead_load_opt.*
* hoist_enters.*
* split_slots.*

This change is necessary as the former design where Slots are users of
the Enter node is problematic in the new optimizer. Furthermore, the new
design is very similar to the Alloc node. This makes it easier in the
future to remove the Slot node altogether and just use the Alloc node
instead.