make API more consitent
The naming of ops and getters of Packs and Variadics and now akin to the
one of Lams and Pis. This resolves this weird overload of Def::arity and
Variadic::arity that now is Variadic::domain. Pack now has Pack::domain
as Lam. This coincidentally fixes a bug in the llvm backend where
exactly this naming was confused. This fixes the fannkuch test again.