Linearize match instructions.
To this end, a new intrinsic was added: Predicated.
Predicated sets the mask under which the target continuation is to be
executed. Diverging control flow can be linearized by computing the
relevant masks and rewiring all braches with the predicated intrinsic as
a means to ensure only relevant lanes are to be written. For now, only
store instructions are effected by the active mask.

Match cases are rewritten by constructing allocas for all values that
have to be merged after all cases are executed. Since storing respects
the current mask, this will ensure that a later load will find the
correct values in this location.