Merges emission of constants and non-constants
Constants are turned into C expressions, whereas non-constants are emitted as part of a basic block.
The previous iteration of this backend made incorrect assumptions about what is a "constant", the method
"is_const_primop" simply made a guess based on the Dep::Param flag, but this guess turns out wrong when non-foldable things like this occur:

mul(2, sizeof(...))

This breaks as emit_constants never expected to see multiplications (it
thinks all of this should have been reduced to an integer, which is
clearly not possible as thorin doesn't know the answer to sizeof).

Rather than duplicate the code responsible of emitting primops for a expr
and BB context, this commit merges the two, asserting when trying to
emit something that needs to be part of a {} block when none is
available.