C: make function locals invisible to other scopes
Certain Defs will end up as function locals, as the notions of Thorin
constants, C constants, expressions vs instructions etc differ between
the two languages. There is no simple mapping.

For instance, Insert _needs_ a basic block to allocate a variable, store
the original value, store the modified component, then return the rest.
If for some reason the same Insert is used in two scopes, we run into
the problem of referencing a local variable from a function into
another. Which of course does not work.

This commit ensures than when we allocate a local variable to
translate a thorin Def, the mapping from Def -> string will be removed
once the scope is finalised, thus guarding against that entire class of
bugs.