refactored the LLVM backend to be less continuation-oriented
Many methods in the LLVM back-end assume a continuation as one of the operands of calls to intrinsics. This is inflexible and brittle - this commit instead have the intrinsics yield LLVM values directly (see `emit_intrinsic`), and then `emit_epilogue` is made responsible for forwarding those values alongside control-flow.

To help in this, we extract a new `emit_call` method: this method is in charge of "making calls", with supplied LLVM values. What is being called can vary: it can be a basic block, a direct-style function, a return parameter, a closure etc ...

`emit_epilogue` works in concert with `emit_call` and is in charge of calling it a second time for direct-style calls, for forwarding the argumnents.

Finally, `emit_epilogue` might need to terminate the basic block if it otherwise ends in a function call.