WIP: evaluate plugins during PE, not after.
This is not very stable for now. Especially the dependency management
can still break at any point. Thorins uses tracking is not very well
maintained, and can't really deal with dependencies that are dropped by
replacing a continuation with a specialized one. This can lead to dead
code being interpreted as a remaining dependency that will only be
removed when the entire world is being rewritten.